Refactored UI code and updated Dockerfile
