+++ management-ui/README.md
@@ -1,22 +1,18 @@
 # Management UI
 
-## Introduction
+This is a full-stack web application based on Python classes and Vanilla JS. It interacts with the 5GMS Application Function via M1 Interface and enables publishing of the M8 json file.
+It can be utilized as the part of overall 5GMS architecture or isolated with 5GMS AF for development purposes.
+Set of provisioning operations implemented in this application is defiend in the TS26.512.
+Backend server is based on FastAPI framework to convert CLI Python methods into web endpoints. Frontend JavaScript `modules` are sending dedicated HTTP requests to the backend, with the required payload generated through the stylized HTML `forms`.
 
-A web-based Graphical User Interface for 5GMS management that uses the [Python library](../python/README.md) to
-interact with the 5GMS Application Function. The source files and the documentation are located in
-the `management-ui` folder of this repository.
 
 ## Installation
 
-There are two ways to run this application. The recommendation is to use the **Docker Compose** service, because
-lightweight container building and activation will solve the entire scope of dependencies and the rest of the
-requirements.
+### Option 1: Docker image
 
-### Option 1. Docker Compose
+To install Docker Compose, follow the official [documentation](https://docs.docker.com/compose/install/).
 
-To install Docker Compose service follow the official [documentation](https://docs.docker.com/compose/install/). Next,
-clone this
-repository:
+Clone 5GMS Application Provider repository:
 
 ```
 cd
@@ -24,40 +20,36 @@ git clone https://github.com/5G-MAG/rt-5gms-application-provider
 cd ~/rt-5gms-application-provider
 ```
 
-Building the Docker image will effectively install all dependencies for the 5GMS Application Function and the Management
-UI:
+Building the Docker image will effectively install all dependencies for the 5GMS Application Function and the Management UI:
 
 ```
 sudo docker-compose build
 ```
 
-Upon successful completion, activate Application Provider with:
-
+Upon successful completion, activate 5GMS Application Provider with:
 ```
 sudo docker-compose up
 ```
 
 Access the module at: `localhost:8000`
 
-### Option 2. Separate installation
+### Option 2: Separate installation
 
-If you prefer to run the 5GMS Application Function separately,
-without setting it up in the Docker environment, you have to build and install it as a local user. For that
-purpose, please follow
+Firstly install 5GMS Application Function following
 this [documentation](https://5g-mag.github.io/Getting-Started/pages/5g-media-streaming/usage/application-function/installation-local-user-5GMSAF.html).
 
-Once installed and built, run the 5GMS Application Function:
-
-```
-~/rt-5gms-application-function/install/bin/open5gs-msafd
-```
-
-Subsequently, install the Python dependencies required for Management UI:
+Subsequently, install the Python dependencies required for the Management UI:
 
 ```
 cd ~/rt-5gms-application-provider
-python3 -m pip install ./python
+sudo python3 -m pip install ./python
 pip3 install -r requirements.txt
+```
+
+Activate 5GMS Application Function:
+
+```
+~/rt-5gms-application-function/install/bin/open5gs-msafd
 ```
 
 Activate the Management UI application:
@@ -70,6 +62,7 @@ uvicorn server:app --reload
 The Management UI will be accessible at port `8000`.
 
 In case that you're receving any kind of web proxy errors, you can bypass that by using empty `http` and `https` proxy flags when running the server:
+
 ```
 cd management-ui/
 http_proxy= https_proxy= uvicorn server:app --reload
@@ -77,18 +70,11 @@ http_proxy= https_proxy= uvicorn server:app --reload
 
 ## Testing
 
-This repository contains CI/CD workflows for building native Docker image, Docker Compose and integration test for
-Management UI application.
+All endpoints are covered with this test, which runs one provisioning cycle and activates all network procedures, checks the responses before deleting all resources.
 
-Automated integration test is written to provide entire provisioning cycle, starting with creation of provisioning
-session, activating all network procedures, and finalizing with deletion of all resources. It effectively conducts
-sequence of HTTP requests to every Management UI web server's endpoint.
-
-Run integration test :
+Run automated test:
 
 ```
 cd ~/rt-5gms-application-provider/management-ui/tests
 pytest integration_test.py
 ```
-
-Please be aware that this procedure is already provided with the repository's CI/CD pipeline.
