+++ management-ui/src/forms/consumptionReporting.html
@@ -1,0 +1,18 @@
+<input id="swal-input1" class="swal2-input" type="number" placeholder="Reporting Interval">
+<input id="swal-input2" class="swal2-input" type="number" placeholder="Sample Percentage">
+<br>
+<br>
+<label for="swal-input3">Location Reporting: </label>
+<br>
+<select id="swal-input3" class="swal2-input">
+  <option value="true">True</option>
+  <option value="false">False</option>
+</select>
+<br>
+<br>
+<label for="swal-input4">Access Reporting: </label>
+<br>
+<select id="swal-input4" class="swal2-input">
+  <option value="true">True</option>
+  <option value="false">False</option>
+</select>
