+++ management-ui/src/forms/dynamicPolicies.html
@@ -1,0 +1,41 @@
+<input id="externalReference" class="swal2-input" type="number" placeholder="External Policy ID" required>
+      
+      <br><br><p>Application Session Context:</p>
+        <input id="sst" class="swal2-input" type="number" placeholder="SST">
+        <input id="sd" class="swal2-input" placeholder="SD">
+        <input id="dnn" class="swal2-input" type="text" placeholder="DNN">
+  
+        <br><br><p font-weight="bold">QoS Specification:</p>
+        <input id="qosReference" class="swal2-input" placeholder="QoS Reference"><br>
+        <br><input id="maxAuthBtrUl" class="swal2-input" type="number" placeholder="Max Auth Btr Ul">
+        <select id="maxAuthBtrUlUnit" class="swal2-input">
+          <option value="bps">Bps</option>
+          <option value="kbps">Kbps</option>
+          <option value="mbps">Mbps</option>
+          <option value="gbps">Gbps</option>
+          <option value="tbps">Tbps</option>
+        </select>
+        <br><input id="maxAuthBtrDl" class="swal2-input" type="number" placeholder="Max Auth Btr Dl">
+        <select id="maxAuthBtrDlUnit" class="swal2-input">
+          <option value="bps">Bps</option>
+          <option value="kbps">Kbps</option>
+          <option value="mbps">Mbps</option>
+          <option value="gbps">Gbps</option>
+          <option value="tbps">Tbps</option>
+        </select>
+        <br>
+        <input id="defPacketLossRateDl" class="swal2-input" placeholder="Def Packet Loss Rate Dl">
+        <input id="defPacketLossRateUl" class="swal2-input" placeholder="Def Packet Loss Rate Ul">
+  
+        <br><br><p>Charging Specification</p>
+        <input id="sponId" class="swal2-input" placeholder="Sponsor ID">
+        <select id="sponStatus" class="swal2-input">
+          <option value="">Select Sponsor Status</option>
+          <option value="SPONSOR_ENABLED">ENABLED</option>
+          <option value="SPONSOR_DISABLED">DISABLED</option>
+        </select>
+        <input id="gpsi" class="swal2-input" placeholder="GPSI">
+  
+  
+        <input id="state" class="swal2-input" placeholder="State">
+        <input id="type" class="swal2-input" placeholder="Type">
