+++ management-ui/src/forms/metricsReporting.html
@@ -1,0 +1,14 @@
+<input id="scheme" class="swal2-input" placeholder="Scheme">
+            <input id="dataNetworkName" class="swal2-input" type='text' placeholder="Data Network Name" required>
+            <input id="reportingInterval" class="swal2-input" type="number" placeholder="Reporting Interval (in seconds)" required>
+            <input id="samplePercentage" class="swal2-input" type="number" placeholder="Sample Percentage (in seconds)" required>
+            <input id="urlFilters" class="swal2-input" placeholder="URL Filters (comma-separated)" required>
+            <input id="samplingPeriod" class="swal2-input" type="number" placeholder="Sampling Period (in seconds)" required>
+            <div>
+            <br><p>Select Metrics to report:</p>
+              <input type="checkbox" id="metric1" value="urn:3GPP:ns:PSS:DASH:QM10#HTTPList"><label for="metric1">HTTP List</label><br>
+              <input type="checkbox" id="metric2" value="urn:3GPP:ns:PSS:DASH:QM10#BufferLevel"><label for="metric2">Buffer Level</label><br>
+              <input type="checkbox" id="metric3" value="urn:3GPP:ns:PSS:DASH:QM10#RepSwitchList"><label for="metric3">Representation Switch List</label><br>
+              <input type="checkbox" id="metric4" value="urn:3GPP:ns:PSS:DASH:QM10#MPDInformation"><label for="metric4">MPD Information</label><br>
+              <input type="checkbox" id="metric5" value="urn:3gpp:metadata:2020:VR:metrics#RenderedViewports"><label for="metric5">Rendered Viewports</label>
+            </div>
