+++ management-ui/src/modules/protocols.js
@@ -1,0 +1,16 @@
+/*
+License: 5G-MAG Public License (v1.0)
+Author: Vuk Stojkovic
+Copyright: (C) Fraunhofer FOKUS
+For full license terms please see the LICENSE file distributed with this
+program. If this file is missing then the license can be retrieved from
+https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
+*/
+
+let operatingUrl = '';
+
+
+export function showProtocols(sessionId) {
+    window.open(`${operatingUrl}show_protocol/${sessionId}`, '_blank');
+  }
+
