+++ management-ui/src/style.css
@@ -309,16 +309,15 @@ button:disabled {
 
 @font-face {
   font-family: 'Poppins ExtraLight';
-  src: url('fonts/Poppins-ExtraLight.ttf') format('truetype');
+  src: url('/src/static/fonts/Poppins-ExtraLight.ttf') format('truetype');
   font-weight: 200;
   font-style: normal;
   font-display: swap;
 }
 
-
 @font-face {
   font-family: 'Poppins Bold';
-  src: url('fonts/Poppins-Bold.ttf') format('truetype');
+  src: url('/src/static/fonts/Poppins-Bold.ttf') format('truetype');
   font-weight: 200;
   font-style: normal;
   font-display: swap;
@@ -326,7 +325,7 @@ button:disabled {
 
 @font-face {
   font-family: 'Poppins Semibold';
-  src: url('fonts/Poppins-SemiBold.ttf') format('truetype');
+  src: url('/src/static/fonts/Poppins-SemiBold.ttf') format('truetype');
   font-weight: 200;
   font-style: normal;
   font-display: swap;
