+++ management-ui/src/templates/index.html
@@ -11,11 +11,11 @@ https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
 <html>
   <head>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
-    <link rel="stylesheet" href="static/style.css">
+    <link rel="stylesheet" href="/src/style.css">
     <link href="https://cdn.jsdelivr.net/npm/sweetalert2@11.7.22/dist/sweetalert2.min.css" rel="stylesheet">
-    <link rel="shortcut icon" type="image/png" href="static/images/favico.png"/>
+    <link rel="shortcut icon" type="image/png" href="src/static/images/favico.png"/>
     <title>5GMS Application Provider</title>
-    <script src="static/m1-user-interface.js"></script>
+    <script type="module" src="/src/provisioning-interface.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.7.22/dist/sweetalert2.all.min.js"></script>
   </head>
   
@@ -23,21 +23,19 @@ https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
 
   <h2 align="center" style="color:white;">5GMS Application Provider</h2>
   <div class="header">
-    <img src="static/images/5gmag.png" alt="5GMAG Logo" class="logo">
+    <img src="src/static/images/5gmag.png" alt="5GMAG Logo" class="logo">
     <br>
     <h4 id="AFStatus" align="center" style="color:black;">Connection with Application Function...</h4>
   </div>
 
-  <div align="left" class="container">
+  <div align="center" class="container">
     <div class="create-session">
         <br>
-        <button onclick="createNewSession()" class="create-session-button">Create Provisioning Session</button><br>
+        <button onclick="createNewSession()" class="create-session-button">Create Provisioning Session</button>
     </div>
 </div>
 
-  
-
-  <table id="m1_table">
+<table id="m1_table">
     <thead>
       <tr>
         <th>Provisioning Session ID</th>
@@ -66,8 +64,8 @@ https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
         For the full license terms, please see the LICENSE file distributed along with 5GMS Reference Tools or retrieve it <a href="https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view">here</a>.
       </p>
       <div class="icon-container">
-        <img src="static/images/fokus.png" alt="FOKUS">
-        <img src="static/images/bbc.png" alt="BBC">
+        <img src="src/static/images/fokus.png" alt="FOKUS">
+        <img src="src/static/images/bbc.png" alt="BBC">
       </div>
     </div>
   </div>
