+++ management-ui/src/utils.js
@@ -1,0 +1,7 @@
+export async function fetchHtmlForm(path) {
+    const response = await fetch(path);
+    if (!response.ok) {
+      throw new Error(`Failed to load HTML from ${path}`);
+    }
+    return response.text();
+  }
